Changed some files to be inside folders (check component and pages) renamed to index.js so it could work. Landing now changes button and routing when signed in vs not signed in added a tweet model, WIUp
