+++ client/src/App/components/Footer/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faTruckMoving, faCopyright  } from '@fortawesome/free-solid-svg-icons'; 
-import './footer.css';
+import './Footer.css';
 
 const Footer = () => {
   return (
