+++ models/index.js
@@ -1,6 +1,7 @@
 module.exports = {
     Vendor: require("./vendor"),
     Menu: require("./menu"),
-    Food: require("./food")
+    Food: require("./food"),
+    Tweet: require("./tweet")
   };
