+++ package.json
@@ -24,10 +24,15 @@
     "mongoose": "^5.7.10",
     "morgan": "^1.9.1",
     "nodemon": "^1.18.7",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
     "react-responsive": "^8.0.1",
     "react-router-dom": "^5.1.2"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "@babel/cli": "^7.7.4",
+    "@babel/core": "^7.7.4"
+  },
   "eslintConfig": {
     "extends": "react-app"
   }
